+++ python/runtime/cudaq/dynamics/CMakeLists.txt
@@ -34,7 +34,6 @@ target_include_directories(nvqir_dynamics_bindings
 target_link_libraries(nvqir_dynamics_bindings PRIVATE
     cudaq-logger
     fmt::fmt-header-only
-    cudaq-mlir-runtime-headers
 )
 # Set output directory for ctest-based python test invocation, which uses cudaq python from the build directory. 
 set_target_properties(nvqir_dynamics_bindings PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/python/cudaq/dynamics)
