+++ python/runtime/interop/PythonCppInterop.cpp
@@ -7,7 +7,7 @@
  ******************************************************************************/
 
 #include "PythonCppInterop.h"
-#include "cudaq.h" // unfortunately, cudaq::get_quake is here at top level
+#include "common/DeviceCodeRegistry.h"
 #include "cudaq/utils/cudaq_utils.h"
 
 cudaq::python::CppPyKernelDecorator::~CppPyKernelDecorator() {
