+++ runtime/common/BraketExecutor.h
@@ -11,7 +11,6 @@
 #include "common/BraketServerHelper.h"
 #include "common/Executor.h"
 #include "common/SampleResult.h"
-#include "cudaq.h"
 #include "cudaq/runtime/logger/logger.h"
 #include <aws/braket/BraketClient.h>
 #include <aws/core/Aws.h>
