+++ runtime/common/DeviceCodeRegistry.h
@@ -51,6 +51,9 @@ void *__cudaq_getRunnableKernel(const std::string &kernelName);
 namespace detail {
 /// Is the kernel `kernelName` registered?
 bool isKernelGenerated(const std::string &kernelName);
+
+/// Is the kernel `kernelName` in library mode?
+bool isLibraryMode(const std::string &kernelName);
 } // namespace detail
 
 /// @brief Given a string kernel name, return the corresponding Quake code
@@ -70,4 +73,19 @@ std::string get_quake_by_name(const std::string &kernelName,
 std::string
 get_quake_by_name(const std::string &kernelName, bool throwException,
                   std::optional<std::string> knownMangledArgs = std::nullopt);
+
+/// Get the name of a plain old function that is marked as a quantum kernel.
+inline std::string get_kernel_function_name(std::string &&name) {
+  return "function_" + std::move(name);
+}
+
+inline std::string get_quake(std::string &&functionName,
+                             const std::string &knownMangledArgs) {
+  return get_quake_by_name(get_kernel_function_name(std::move(functionName)),
+                           knownMangledArgs);
+}
+
+typedef std::size_t (*KernelArgsCreator)(void **, void **);
+KernelArgsCreator getArgsCreator(const std::string &kernelName);
+
 } // namespace cudaq
