+++ runtime/cudaq/algorithms/evolve.cpp
@@ -8,13 +8,13 @@
 
 #include "common/AnalogHamiltonian.h"
 #include "common/EvolveResult.h"
+#include "cudaq.h"
 #include "cudaq/operators.h"
 #include "cudaq/runtime/logger/logger.h"
 #include "cudaq/schedule.h"
 #include <random>
 #include <sstream>
 #include <string>
-
 namespace cudaq::__internal__ {
 
 evolve_result evolveSingle(const cudaq::rydberg_hamiltonian &hamiltonian,
