+++ runtime/cudaq/platform/qpu_state.h
@@ -9,7 +9,6 @@
 #pragma once
 
 #include "common/SimulationState.h"
-#include "cudaq.h"
 #include "cudaq/utils/cudaq_utils.h"
 #include <type_traits>
 #include <utility>
