+++ runtime/nvqir/cudensitymat/CuDensityMatContext.cpp
@@ -9,7 +9,7 @@
 #include "CuDensityMatContext.h"
 #include "CuDensityMatErrorHandling.h"
 #include "CuDensityMatUtils.h"
-#include "cudaq.h"
+#include "cudaq/cudaq_mpi.h"
 #include "cudaq/distributed/mpi_plugin.h"
 #include "cudaq/runtime/logger/logger.h"
 #include <memory>
